+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Added a documentation based on the Filter Factory documentation.
 - Added a help button which opens the documentation on the web. (*)
 - For 32/64-bit cross-build, there is now no requirement that both plugin files must be installed anymore.
+- Windows 95/98/Me: If UNICOWS.DLL is missing (required for building a standalone filter), it will be automatically installed for you.
 
 (*) This bug/solution was tested on Windows but needs to be verified and/or implemented on Mac.